pyproject.toml
[project]
name = "pointsball"
version = "0.1.0"
description = "Automated Fantasy Premier League team management"
requires-python = ">=3.10,<3.12"
dependencies = [
    "polars>=1.0",
    "fpl>=0.6.35",
    "aiohttp>=3.9.3",
    "praw>=7.7.1",
    "python-dotenv>=1.0",
    "typer>=0.24.1",
    "scikit-learn>=1.7.2",
    "ortools>=9.15.6755",
    "requests>=2.32.5",
]

[project.scripts]
pointsball = "pointsball.__main__:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.2.1",
    "black>=24.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
skip-string-normalization = true
line-length = 120

[tool.ruff]
line-length = 120
exclude = ["jupyter_notebook_config.py"]
lint.select = [
    "E",
    "W",
    "F",
    "I",
    "C",
    "B",
]
lint.ignore = [
    "E501",
    "C901",
    "B008",  # typer.Option in function defaults is the intended Typer pattern
]

[tool.ruff.lint.isort]
order-by-type = true
relative-imports-order = "closest-to-furthest"
extra-standard-library = ["typing"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]
known-first-party = []

[dependency-groups]
dev = [
    "black>=26.3.1",
    "ruff>=0.15.6",
]
